requirements.txt
numpy>=1.18.1
hyperopt>=0.2.3
lightgbm>=2.3.1
pandas>=0.25.3
shap>=0.34.0
matplotlib>=3.1.2
seaborn>=0.9.0
xgboost>=0.90
catboost>=0.20.2
category_encoders>=2.1.0
sphinx==2.4.1
sphinx-argparse==0.2.5
sphinx-autodoc-typehints==1.10.3
nbsphinx==0.5.1
recommonmark==0.6.0
sphinx_copybutton==0.2.8
RinohType==0.3.1
pypng==0.0.20
pytest==5.3.5
pyyaml==5.3